SMS-55: Create API get Employee

+++ src/main/java/com/sms/be/model/Employee.java
@@ -32,7 +32,7 @@ public class Employee extends BaseEntity {
     private Long salary;
 
     @Column
-    private Double avgRating;
+    private Long avgRating;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private Salon salon;
